+++ osidb/models/ps_product.py
@@ -2,8 +2,6 @@ import uuid
 
 from django.db import models
 
-from osidb.helpers import deprecate_field
-
 
 class PsProduct(models.Model):
     # internal primary key
@@ -14,9 +12,6 @@ class PsProduct(models.Model):
 
     # name of the product
     name = models.CharField(max_length=100)
-
-    # team responsible for the product
-    team = deprecate_field(models.CharField(max_length=50))
 
     # the business unit to which the product belongs
     business_unit = models.CharField(max_length=50)